changing uint_8 to int

+++ firmware/examples/HT16K33/HT16K33.ino
@@ -19,13 +19,12 @@
   BSD license, all text above must be included in any redistribution
  ****************************************************/
 
-#include <Wire.h>
 #include "Adafruit_LEDBackpack.h"
 #include "Adafruit_GFX.h"
 
 Adafruit_LEDBackpack matrix = Adafruit_LEDBackpack();
 
-uint8_t counter = 0;
+int counter = 0;
 
 void setup() {
   Serial.begin(9600);
@@ -37,7 +36,7 @@ void setup() {
 void loop() {
   // paint one LED per row. The HT16K33 internal memory looks like
   // a 8x16 bit matrix (8 rows, 16 columns)
-  for (uint8_t i=0; i<8; i++) {
+  for (int i=0; i<8; i++) {
     // draw a diagonal row of pixels
     matrix.displaybuffer[i] = _BV((counter+i) % 16) | _BV((counter+i+8) % 16)  ;
   }
